artifacts/submission/aurabeam_source_code/requirements.txt
numpy
opencv-python
pyserial
ultralytics
